Make the equivalence tests hold on a GPU
The suite's first run on a GPU - the 4090 on goalmisgen-h1, after the noise
programme's tests were written against the Surface's CPU - failed four tests,
all of the same species: two orderings of the same computation asserted to
agree. Three were TF32: on Ampere and newer, JAX lowers float32 matmuls to
tensor cores carrying ~3 decimal digits, so transcription checks that agree to
1e-7 on any CPU disagree at 1e-3. A conftest pins the tests, and only the
tests, to full-precision matmuls.

The fourth was the remat cosine, asserted to 1e-12 on a float32 dot product -
which float32 accumulation cannot certify and which held on CPU only by luck
of reassociation. The test now accumulates in float64; the gradients it
compares are unchanged.

+++ tests/test_offline_model.py
@@ -327,7 +327,10 @@ def test_remat_gradients_differ_only_by_float32_reassociation() -> None:
 
     def flat_grad(model):
         grads = jax.grad(lambda p: model.apply(p, observations, actions)[0].sum())(params)
-        return np.concatenate([np.asarray(leaf).ravel() for leaf in jax.tree_util.tree_leaves(grads)])
+        # float64: the cosine below is asserted to 1e-12, which the gradients'
+        # own float32 accumulation cannot certify - it held on CPU only by luck
+        # of reassociation, and a GPU's different summation order exposed it.
+        return np.concatenate([np.asarray(leaf).ravel() for leaf in jax.tree_util.tree_leaves(grads)]).astype(np.float64)
 
     with_remat, without = flat_grad(on), flat_grad(off)
     assert np.linalg.norm(with_remat - without) / np.linalg.norm(without) < 1e-5
